Reorganising the admin code to make it more maintainable
It was a bad idea to have playlist-editor and filelist-editor as
independent applications. When managing and not narrating, it
makes more sense that they are a single, integrated application.
Thus, move all the logic to a new component <story-editor> and
re-wire everything.

While at it, try to be more consistent and robust about everything.