UI tweaks
Remove the story rename functionality and button. Instead, show
the story title in the story editing interface and add a button
to rename the story.

Moreover, add a "Narrate" button directly in the story list, so
that it's not necessary to go to the edit interface in order to
narrate a story.