Refactor tools to return immutable operations
Instead of receiving the graphic context and drawing directly on
it, make tools return a list of operations to do on the canvas, and
apply them from the MapDiscoverer itself. This makes it easier to
unit test the tools, and to add other, more abstract operations
that are not necessarily drawing on a canvas.

While at it, generalise the single penSize number to a properties
object.